4.0 version change
This is more symbolic than anything. Sure, rpg is now a public command, and that was a big thing, but the game isn't exactly finished, and not much else was changed significantly.
It's more about how the bot has developed so far through the past few months and how I think it's going to develop in the future. I feel more comfortable with the project now, thus the documentation website linked in the readme.