Fixed a couple things, seemingly, and added a scheduled restart
I don't think the scheduling service was even working my dude. Hopefully this doesn't break anything.
I still don't know why exactly it is that very rarely eating a ghost kills you. Hopefully it's not a race condition, in which case I'd probably need a masters degree in computer science to fix it